rename: JwtLogoutHandler -> CustomLogoutHandler
